CXP-1333: Refactor product selection formatting to pqb filter
