+++ components/catalogs/back/src/Infrastructure/Persistence/Catalog/Product/GetProductIdentifiersQuery.php
@@ -6,6 +6,7 @@ namespace Akeneo\Catalogs\Infrastructure\Persistence\Catalog\Product;
 
 use Akeneo\Catalogs\Application\Persistence\Catalog\Product\GetProductIdentifiersQueryInterface;
 use Akeneo\Catalogs\Domain\Catalog;
+use Akeneo\Catalogs\Infrastructure\Service\FormatProductSelectionCriteria;
 use Akeneo\Pim\Enrichment\Bundle\Elasticsearch\IdentifierResult;
 use Akeneo\Pim\Enrichment\Component\Product\Query\ProductQueryBuilderFactoryInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Query\Sorter\Directions;
@@ -33,7 +34,7 @@ final class GetProductIdentifiersQuery implements GetProductIdentifiersQueryInte
     public function execute(Catalog $catalog, ?string $searchAfter = null, int $limit = 100): array
     {
         $pqbOptions = [
-            'filters' => $this->getFilters($catalog),
+            'filters' => FormatProductSelectionCriteria::toPQBFilters($catalog->getProductSelectionCriteria()),
             'limit' => $limit,
         ];
 
@@ -75,33 +76,6 @@ final class GetProductIdentifiersQuery implements GetProductIdentifiersQueryInte
         }
 
         return (string) $identifier;
-    }
-
-    /**
-     * @return array<mixed>
-     */
-    private function getFilters(Catalog $catalog): array
-    {
-        $filters = [];
-        /** @var array<array-key, array{field: string, operator: string, value?: mixed, scope?: string|null, locale?: string|null}> $productSelectionCriteria */
-        $productSelectionCriteria = $catalog->getProductSelectionCriteria();
-        foreach ($productSelectionCriteria as $criterion) {
-            $filter = $criterion;
-
-            if (isset($criterion['scope'])) {
-                $filter['context']['scope'] = $criterion['scope'];
-            }
-
-            if (isset($criterion['locale'])) {
-                $filter['context']['locale'] = $criterion['locale'];
-            }
-
-            unset($filter['scope'], $filter['locale']);
-
-            $filters[] = $filter;
-        }
-
-        return $filters;
     }
 
     private function getUuidFromIdentifierResult(string $esId): string
