+++ components/catalogs/back/src/Infrastructure/Service/FormatProductSelectionCriteria.php
@@ -2,14 +2,14 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Catalogs\Infrastructure\Persistence\Catalog\Product;
+namespace Akeneo\Catalogs\Infrastructure\Service;
 
 /**
  * @author    Willy Mesnage <willy.mesnage@akeneo.com>
  * @copyright 2022 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-final class GetPQBFilters
+final class FormatProductSelectionCriteria
 {
     /**
      * @param array<
@@ -30,7 +30,7 @@ final class GetPQBFilters
      *     }
      * > $productSelectionCriteria
      */
-    public static function fromProductSelectionCriteria(array $productSelectionCriteria): array
+    public static function toPQBFilters(array $productSelectionCriteria): array
     {
         $filters = [];
         foreach ($productSelectionCriteria as $criterion) {
